reorg aztec.js
