+++ yarn-project/aztec.js/src/account_manager/deploy_account_method.ts
@@ -9,9 +9,9 @@ import type { PublicKeys } from '@aztec/stdlib/keys';
 
 import type { AuthWitnessProvider } from '../account/interface.js';
 import type { Wallet } from '../account/wallet.js';
-import { Contract } from '../api/contract/contract.js';
-import { DeployMethod, type DeployOptions } from '../api/contract/deploy_method.js';
 import type { ExecutionRequestInit } from '../api/entrypoint.js';
+import { Contract } from '../contract/contract.js';
+import { DeployMethod, type DeployOptions } from '../contract/deploy_method.js';
 import { EntrypointPayload, computeCombinedPayloadHash } from '../entrypoint/payload.js';
 
 /**
