+++ yarn-project/aztec.js/src/account_manager/deploy_account_sent_tx.ts
@@ -3,7 +3,7 @@ import type { PXE } from '@aztec/stdlib/interfaces/client';
 import type { TxHash, TxReceipt } from '@aztec/stdlib/tx';
 
 import type { Wallet } from '../account/index.js';
-import { DefaultWaitOpts, SentTx, type WaitOpts } from '../api/contract/sent_tx.js';
+import { DefaultWaitOpts, SentTx, type WaitOpts } from '../contract/sent_tx.js';
 
 /** Extends a transaction receipt with a wallet instance for the newly deployed contract. */
 export type DeployAccountTxReceipt = FieldsOf<TxReceipt> & {
