+++ yarn-project/aztec.js/src/account_manager/index.ts
@@ -7,9 +7,9 @@ import { deriveKeys } from '@aztec/stdlib/keys';
 import type { AccountContract } from '../account/contract.js';
 import type { Salt, Wallet } from '../account/index.js';
 import type { AccountInterface } from '../account/interface.js';
-import { Contract } from '../api/contract/contract.js';
-import { DeployMethod, type DeployOptions } from '../api/contract/deploy_method.js';
-import { DefaultWaitOpts, type WaitOpts } from '../api/contract/sent_tx.js';
+import { Contract } from '../contract/contract.js';
+import { DeployMethod, type DeployOptions } from '../contract/deploy_method.js';
+import { DefaultWaitOpts, type WaitOpts } from '../contract/sent_tx.js';
 import { DefaultMultiCallEntrypoint } from '../entrypoint/default_multi_call_entrypoint.js';
 import { AccountWalletWithSecretKey, SignerlessWallet } from '../wallet/index.js';
 import { DeployAccountMethod } from './deploy_account_method.js';
