+++ yarn-project/aztec.js/src/api/deployment.ts
@@ -1,3 +1,4 @@
 export { registerContractClass } from '../deployment/register_class.js';
 export { broadcastPrivateFunction, broadcastUnconstrainedFunction } from '../deployment/broadcast_function.js';
 export { deployInstance } from '../deployment/deploy_instance.js';
+export { ContractDeployer } from '../deployment/contract_deployer.js';
