+++ yarn-project/aztec.js/src/api/interfaces.ts
