+++ yarn-project/aztec.js/src/api/rpc.ts
@@ -1,0 +1,2 @@
+export { createPXEClient } from '../rpc_clients/pxe_client.js';
+export { createCompatibleClient } from '../rpc_clients/node/index.js';
