--- yarn-project/aztec.js/src/api/test/index.ts
@@ -1,5 +1,0 @@
-export { AnvilTestWatcher } from './anvil_test_watcher.js';
-export { RollupCheatCodes } from './rollup_cheat_codes.js';
-export { AztecCheatCodes } from './aztec_cheat_codes.js';
-export { EthCheatCodes } from '@aztec/ethereum/eth-cheatcodes';
-export { CheatCodes } from './cheat_codes.js';
