+++ yarn-project/aztec.js/src/api/testing.ts
@@ -1,0 +1,5 @@
+export { AnvilTestWatcher } from '../test/anvil_test_watcher.js';
+export { RollupCheatCodes } from '../test/rollup_cheat_codes.js';
+export { AztecCheatCodes } from '../test/aztec_cheat_codes.js';
+export { EthCheatCodes } from '@aztec/ethereum/eth-cheatcodes';
+export { CheatCodes } from '../test/cheat_codes.js';
