+++ yarn-project/aztec.js/src/api/utils.ts
@@ -1,0 +1,21 @@
+export { generatePublicKey } from '../utils/pub_key.js';
+export {
+  type AztecAddressLike,
+  type EthAddressLike,
+  type EventSelectorLike,
+  type FieldLike,
+  type FunctionSelectorLike,
+  type U128Like,
+  type WrappedFieldLike,
+} from '../utils/abi_types.js';
+export {
+  computeAuthWitMessageHash,
+  computeInnerAuthWitHash,
+  computeInnerAuthWitHashFromAction,
+  type IntentAction,
+  type IntentInnerHash,
+} from '../utils/authwit.js';
+export { waitForPXE } from '../utils/pxe.js';
+export { waitForNode, createAztecNodeClient, type AztecNode } from '../utils/node.js';
+export { getFeeJuiceBalance } from '../utils/fee_juice.js';
+export { readFieldCompressedString } from '../utils/field_compressed_string.js';
