--- yarn-project/aztec.js/src/api/utils/index.ts
@@ -1,21 +1,0 @@
-export { generatePublicKey } from './pub_key.js';
-export {
-  type AztecAddressLike,
-  type EthAddressLike,
-  type EventSelectorLike,
-  type FieldLike,
-  type FunctionSelectorLike,
-  type U128Like,
-  type WrappedFieldLike,
-} from './abi_types.js';
-export {
-  computeAuthWitMessageHash,
-  computeInnerAuthWitHash,
-  computeInnerAuthWitHashFromAction,
-  type IntentAction,
-  type IntentInnerHash,
-} from './authwit.js';
-export { waitForPXE } from './pxe.js';
-export { waitForNode, createAztecNodeClient, type AztecNode } from './node.js';
-export { getFeeJuiceBalance } from './fee_juice.js';
-export { readFieldCompressedString } from './field_compressed_string.js';
