+++ yarn-project/aztec.js/src/contract/checker.test.ts
