+++ yarn-project/aztec.js/src/contract/contract_base.ts
@@ -7,7 +7,7 @@ import {
 } from '@aztec/stdlib/abi';
 import { type ContractInstanceWithAddress, computePartialAddress } from '@aztec/stdlib/contract';
 
-import type { Wallet } from '../../account/index.js';
+import type { Wallet } from '../account/wallet.js';
 import { ContractFunctionInteraction } from './contract_function_interaction.js';
 
 /**
