+++ yarn-project/aztec.js/src/contract/deploy_method.ts
@@ -11,10 +11,10 @@ import type { GasSettings } from '@aztec/stdlib/gas';
 import type { PublicKeys } from '@aztec/stdlib/keys';
 import type { Capsule, TxExecutionRequest } from '@aztec/stdlib/tx';
 
-import type { Wallet } from '../../account/index.js';
-import { deployInstance } from '../../deployment/deploy_instance.js';
-import { registerContractClass } from '../../deployment/register_class.js';
-import type { ExecutionRequestInit } from '../../entrypoint/entrypoint.js';
+import type { Wallet } from '../account/wallet.js';
+import { deployInstance } from '../deployment/deploy_instance.js';
+import { registerContractClass } from '../deployment/register_class.js';
+import type { ExecutionRequestInit } from '../entrypoint/entrypoint.js';
 import { BaseContractInteraction, type SendMethodOptions } from './base_contract_interaction.js';
 import type { Contract } from './contract.js';
 import type { ContractBase } from './contract_base.js';
