+++ yarn-project/aztec.js/src/deployment/contract_deployer.ts
@@ -3,8 +3,8 @@ import { AztecAddress } from '@aztec/stdlib/aztec-address';
 import { PublicKeys } from '@aztec/stdlib/keys';
 
 import type { Wallet } from '../account/wallet.js';
-import { Contract } from '../api/contract/contract.js';
-import { DeployMethod } from '../api/contract/deploy_method.js';
+import { Contract } from '../contract/contract.js';
+import { DeployMethod } from '../contract/deploy_method.js';
 
 /**
  * A class for deploying contract.
