--- yarn-project/aztec.js/src/deployment/index.ts
@@ -1,1 +1,0 @@
-export * from './contract_deployer.js';
