+++ yarn-project/aztec.js/src/deployment/register_class.ts
@@ -5,8 +5,8 @@ import { type ContractArtifact, bufferAsFields } from '@aztec/stdlib/abi';
 import { getContractClassFromArtifact } from '@aztec/stdlib/contract';
 import { Capsule } from '@aztec/stdlib/tx';
 
-import type { ContractFunctionInteraction } from '../api/contract/contract_function_interaction.js';
-import { getRegistererContract } from '../api/contract/protocol_contracts.js';
+import type { ContractFunctionInteraction } from '../contract/contract_function_interaction.js';
+import { getRegistererContract } from '../contract/protocol_contracts.js';
 import type { Wallet } from '../wallet/index.js';
 
 const defaultEmitPublicBytecode =
