+++ yarn-project/aztec.js/src/ethereum/l1_contracts.ts
@@ -1,7 +1,7 @@
 import type { L1ContractAddresses } from '@aztec/ethereum/l1-contract-addresses';
 import { retryUntil } from '@aztec/foundation/retry';
 
-import { createPXEClient } from '../../rpc_clients/index.js';
+import { createPXEClient } from '../rpc_clients/pxe_client.js';
 
 export const getL1ContractAddresses = async (url: string): Promise<L1ContractAddresses> => {
   const pxeClient = createPXEClient(url, {});
