+++ yarn-project/aztec.js/src/fee/private_fee_payment_method.ts
@@ -4,7 +4,7 @@ import type { AztecAddress } from '@aztec/stdlib/aztec-address';
 import type { GasSettings } from '@aztec/stdlib/gas';
 
 import type { Wallet } from '../account/wallet.js';
-import { ContractFunctionInteraction } from '../api/contract/contract_function_interaction.js';
+import { ContractFunctionInteraction } from '../contract/contract_function_interaction.js';
 import { SignerlessWallet } from '../wallet/signerless_wallet.js';
 import type { FeePaymentMethod } from './fee_payment_method.js';
