--- yarn-project/aztec.js/src/rpc_clients/index.ts
@@ -1,2 +1,0 @@
-export * from './pxe_client.js';
-export * from './node/index.js';
