+++ yarn-project/aztec.js/src/utils/authwit.ts
