+++ yarn-project/aztec.js/src/utils/fee_juice.ts
@@ -2,8 +2,7 @@ import { Fr } from '@aztec/foundation/fields';
 import { ProtocolContractAddress } from '@aztec/protocol-contracts';
 import type { AztecAddress } from '@aztec/stdlib/aztec-address';
 import { deriveStorageSlotInMap } from '@aztec/stdlib/hash';
-
-import type { PXE } from '../../api/interfaces/pxe.js';
+import type { PXE } from '@aztec/stdlib/interfaces/client';
 
 /**
  * Returns the owner's fee juice balance.
