+++ yarn-project/aztec.js/src/utils/field_compressed_string.ts
