+++ yarn-project/aztec.js/src/utils/node.ts
