+++ yarn-project/aztec.js/src/utils/pub_key.ts
