+++ yarn-project/aztec.js/src/wallet/base_wallet.ts
@@ -29,8 +29,8 @@ import type {
 } from '@aztec/stdlib/tx';
 
 import type { Wallet } from '../account/wallet.js';
-import type { IntentAction, IntentInnerHash } from '../api/utils/authwit.js';
 import type { ExecutionRequestInit } from '../entrypoint/entrypoint.js';
+import type { IntentAction, IntentInnerHash } from '../utils/authwit.js';
 
 /**
  * A base class for Wallet implementations
